+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.3
 - Added config option for changing the temporal namespace
 - Added config options to allow customization of default retry options for workflows and activities
+- Added `--scoped` option to `make` commands to allow generating Workflow/Activity inside a scoped namespace
 
 ## v0.2
 - Added testing helpers: activity/workflows mocks, dispatches assertions, automatic test server and worker